+++ p2pRes/src/p2pRes/protocol/response/AskForBlock.java
@@ -2,14 +2,14 @@ package p2pRes.protocol.response;
 
 public class AskForBlock extends ProtocolResponse {
 
-	private final long blockNumber;
+	private final int blockNumber;
 	
-	public AskForBlock(long blockNumber) {
+	public AskForBlock(int blockNumber) {
 		super(Command.ASK_FOR_BLOCK);
 		this.blockNumber = blockNumber;
 	}
 
-	public long getBlockNumber() {
+	public int getBlockNumber() {
 		return blockNumber;
 	}
 }
